+++ app/Http/Controllers/ControllerMasterCustomer.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Mcust;
+use App\Models\Tposh;
 use Illuminate\Http\Request;
 
 class ControllerMasterCustomer extends Controller
@@ -49,7 +50,27 @@ class ControllerMasterCustomer extends Controller
     }
 
     public function delete(Mcust $mcust){
-        Mcust::find($mcust->id)->delete();
-        return redirect()->route('mcust');
+        // Mcust::find($mcust->id)->delete();
+        // return redirect()->route('mcust');
+
+        $item = Mcust::where('id','=',$mcust->id)->get();
+        foreach($item as $data){
+            $itemname = $data->name;
+        }
+
+        $pos = Tposh::select('code_mcust')->where('code_mcust', 'like', '%' . $itemname . '%')->whereNull('deleted_at')->first();
+        // dd($pos);
+        
+        $havetrans = 0;
+        if($pos != null ){
+            $havetrans = 1;
+        }
+
+        if($havetrans == 0){
+            Mcust::find($mcust->id)->delete();
+            return redirect()->route('mcust')->with('success','Data Berhasil Di Hapus');
+        }else{
+            return redirect()->route('mcust')->with('error','Tidak dapat menghapus data, karena data ada di dalam transaksi lainnya');
+        }
     }
 }
