+++ app/Http/Controllers/ControllerMasterMataUang.php
@@ -3,6 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Mmatauang;
+use App\Models\Tbayaropsh;
+use App\Models\Tjurnalvouchh;
+use App\Models\Tpembelianh;
+use App\Models\Tposh;
 use Illuminate\Http\Request;
 
 class ControllerMasterMataUang extends Controller
@@ -45,8 +49,29 @@ class ControllerMasterMataUang extends Controller
     }
 
     public function delete(Mmatauang $mmatauang){
-        // dd($muom);
-        Mmatauang::find($mmatauang->id)->delete();
-        return redirect()->route('mmatauang');
+        // Mmatauang::find($mmatauang->id)->delete();
+        // return redirect()->route('mmatauang');
+
+        $item = Mmatauang::where('id','=',$mmatauang->id)->get();
+        foreach($item as $data){
+            $itemname = $data->name;
+        }
+        $pembelian = Tpembelianh::select('mata_uang')->where('mata_uang', 'like', '%' . $itemname . '%')->whereNull('deleted_at')->first();
+        $pos = Tposh::select('mata_uang')->where('mata_uang', 'like', '%' . $itemname . '%')->whereNull('deleted_at')->first();
+        $bayarops = Tbayaropsh::select('mata_uang')->where('mata_uang', 'like', '%' . $itemname . '%')->whereNull('deleted_at')->first();
+        $jurnal_vouch = Tjurnalvouchh::select('mata_uang')->where('mata_uang', 'like', '%' . $itemname . '%')->whereNull('deleted_at')->first();
+        // dd($pembelian);
+        
+        $havetrans = 0;
+        if($pembelian != null || $pos != null || $bayarops != null || $jurnal_vouch != null){
+            $havetrans = 1;
+        }
+
+        if($havetrans == 0){
+            Mmatauang::find($mmatauang->id)->delete();
+            return redirect()->route('mmatauang')->with('success','Data Berhasil Di Hapus');
+        }else{
+            return redirect()->route('mmatauang')->with('error','Tidak dapat menghapus data, karena data ada di dalam transaksi lainnya');
+        }
     }
 }
